Rename spark310 to spark311
