+++ docs/get-started/getting-started-on-prem.md
@@ -433,7 +433,7 @@ that matches your Spark version. Currently we support
    - Spark 3.0.0 (com.nvidia.spark.rapids.spark300.RapidsShuffleManager) 
    - Spark 3.0.1 (com.nvidia.spark.rapids.spark301.RapidsShuffleManager) 
    - Spark 3.0.2 (com.nvidia.spark.rapids.spark302.RapidsShuffleManager) 
-   - Spark 3.1.0 (com.nvidia.spark.rapids.spark310.RapidsShuffleManager)
+   - Spark 3.1.1 (com.nvidia.spark.rapids.spark311.RapidsShuffleManager)
 
 ```shell
 ...
