+++ shims/pom.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-  Copyright (c) 2020, NVIDIA CORPORATION.
+  Copyright (c) 2020-2021, NVIDIA CORPORATION.
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -46,7 +46,7 @@
             </activation>
             <modules>
                 <module>spark302</module>
-                <module>spark310</module>
+                <module>spark311</module>
             </modules>
         </profile>
     </profiles>
