+++ shims/spark311/src/main/resources/META-INF/services/com.nvidia.spark.rapids.SparkShimServiceProvider
@@ -1,0 +1,1 @@
+com.nvidia.spark.rapids.shims.spark311.SparkShimServiceProvider
