+++ shims/spark311/src/main/scala/org/apache/spark/sql/execution/datasources/parquet/rapids/ParquetMaterializer.scala
