+++ tests-spark310+/README.md
@@ -14,9 +14,9 @@ Example:
 As of writing of this document this module contains tests that are strongly tied to classes in 
 Spark-3.1.0-SNAPSHOT.
   
-These tests can be executed by choosing profile `spark310tests` like so, 
+These tests can be executed by choosing profile `spark311tests` like so, 
   
-`mvn -Pspark310tests -wildcardSuites=<testname wildcard>`
+`mvn -Pspark311tests -wildcardSuites=<testname wildcard>`
 
 For a more comprehensive overview of tests in Rapids Accelerator please refer to the following
