+++ tests/pom.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
-  Copyright (c) 2020, NVIDIA CORPORATION.
+  Copyright (c) 2020-2021, NVIDIA CORPORATION.
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -53,9 +53,9 @@
             </properties>
         </profile>
         <profile>
-            <id>spark310tests</id>
+            <id>spark311tests</id>
             <properties>
-                <spark.test.version>3.1.0-SNAPSHOT</spark.test.version>
+                <spark.test.version>3.1.1-SNAPSHOT</spark.test.version>
             </properties>
         </profile>
     </profiles>
